Fix ordering of latitudes and longitudes
The coordinates were used the wrong way around (lon, lat). This may or
may not have cause issues when calculating distances, but for consistency
sakes they are now presented the other way around (lat,lon). Also made
the location attribute a dictionary instead of a list to avoid any confusions.